+++ aqueduct/lib/src/db/query/query.dart
@@ -56,8 +56,7 @@ abstract class Query<InstanceType extends ManagedObject> {
   /// This is equivalent to creating a [Query], assigning [object] to [values], and invoking [insert].
   static Future<T> insertObject<T extends ManagedObject>(
       ManagedContext context, T object) {
-    final query = Query<T>(context)..values = object;
-    return query.insert();
+    return context.insertObject(object);
   }
 
   /// Inserts each object in [objects] into the database managed by [context] in a single transaction.
@@ -65,7 +64,7 @@ abstract class Query<InstanceType extends ManagedObject> {
   /// This currently has no Query instance equivalent
   static Future<List<T>> insertObjects<T extends ManagedObject>(
       ManagedContext context, List<T> objects) async {
-    return context.transaction((transitionCtx) => Future.wait(objects.map((o) => insertObject(transitionCtx, o))));
+    return context.insertObjects(objects);
   }
 
   /// Configures this instance to fetch a relationship property identified by [object] or [set].