+++ aqueduct/CHANGELOG.md
@@ -4,10 +4,10 @@
 - Deprecates `AuthCodeController` in favor of `AuthRedirectController`.
 - Improves speed of many database CLI commands
 - Improves error messaging of the CLI; no longer includes stack trace for expected errors.
-- Adds `Query.insertObjects` to insert a list of managed objects.
 - Allows self-referencing and cyclical relationships between managed objects
 - Autoincrementing managed object properties are no longer consumed in `ManagedObject.readFromMap`.
 - Fixes bug where ManagedObjects cannot have mixins
+- Adds `ManagedContext.insertObject`, `ManagedContext.insertObjects` and `ManagedContext.fetchObjectWithID`.
 
 ## 3.0.2